Added a list of objects to be counted.
Melter and Presser prefabs moved into Resources folder to be able to be instantiated in scene.

Added see through area planes for future development into areas in the game.

Object Information scripts to set and gather data on the objects in the scene have been added.
 - Maintenance Cost
 - Operational Cost
 - Object Type, e.g. MELTER.

Added an object Manager to split components as to stop bloating in one area.

Most of it is hard coded at the moment. But just proof of concept.

Grid script now instantiates both prefabs and add to a information list on the objects in the scene (HARD CODED).

Home improvement:
Input script
 - Instead of declaring a variable and then using it later on, it's now declared at usage.

 - Economy script edited to use functions (A singular function could be used instead and just use -value instead of two functions).