Added a 'quarterly' earnings menu (Not functional)
After a certain amount of time has passed a menu showing the 'quarterly' results for the company will be shown.

Time is then paused (will probably have to have a look into how to do this more efficiently when adding more stuff.)

After this time period has passed the operational cost is taken away from the budget and profit is added to budget.

Created a UIScript to create UI functions within, like continuing the game and button functions.
Home Improvements:
 - Added collision boxes around prefabs just for testing ray casting.
 - Added debug logs for economy.
 - Changed spacing between grid system to be floats to get smaller increments between squares.
 - Added some getters to the object information gatherer script

For the most part the code implemented is just for  proof of concept and will need re-aranged.