Added multiple grid system.
Created different default grid objects for start to differentiate grids, FOR NOW.

Changed stat menu sell button to use sell function in the UI script now rather than a defined grid script due to now using multiple and need to check which one is being used.

Added which zone is currently active at bottom of screen for testing.

Updated Grid script to now include the starting locations, so we can position the grids in the scene properly.

Pressing space will toggle between the two zones.

Created a zone decider script to use to define which grid scrip to use when interacting with the object and will also be used to change the active zone when switching between them in-game.