Added health/construction time
Added variables needed for each object/machine.

Added a toggle to determine if the object is a machine or not to deteriorate the machine health over time.

Updated marketplace buy values for the machines to say current purchase values.

Created a co routine for the construction time to be waited for until the object gets placed in scene.

Took out the spacebar input to toggle between current zones.

Update the stat menu when an object is clicked on to show the machines health.

If the machines reached 0, the machine is destroyed.