Updated production grid to include conveyor belts and position of assets.
The production area now has a process of production. Where a stockpile of the output from each machine can be used for the next machine and so on.

A delivery value is used for the start and a shipping value is used for the end.

Economy is affected by the buying/selling of product.

Checks in place for the delivery and the shipping.

Added an on destroy method to each object to update the production input/output values used for the process.

Commented out the screen boundary controls and added in the scroll wheel zoom function.

Gave some values to the prefabs input/output and added a process number for the production.